requirements_dev.txt
flake8==3.9.2
tox==3.24.3
pytest-cov
mypy